requirements.txt
aiosmtpd==1.4.2
atpublic==3.0.1
attrs==22.1.0
prettytable==3.3.0
python-dotenv==0.20.0
wcwidth==0.2.5
